Rough refactor to display UI using child services
